cherrypick

+++ src/representative_subset_calculator/kernel_matrix/kernel_matrix.h
@@ -11,17 +11,17 @@ class KernelMatrix {
      * Needs to be parallel safe for row based indexes, otherwise race conditions will occur
      *  during getDiagonals()
      */
-    virtual double get(size_t j, size_t i) = 0;
+    virtual float get(size_t j, size_t i) = 0;
 
     virtual size_t size() = 0;
 
-    double getCoverage() {
+    float getCoverage() {
         return KernelMatrix::getCoverage(this->getDiagonals());
     }
 
-    std::vector<double> getDiagonals() {
+    std::vector<float> getDiagonals() {
         const size_t s = this->size();
-        std::vector<double> res(s);
+        std::vector<float> res(s);
         
         #pragma omp parallel for
         for (size_t index = 0; index < s; index++) {
@@ -31,17 +31,16 @@ class KernelMatrix {
         return move(res);
     }
 
-    static double getDotProduct(const std::vector<double> &a, const std::vector<double> &b) {
-        double res = 0;
+    static float getDotProduct(const std::vector<float> &a, const std::vector<float> &b) {
+        float res = 0;
         for (size_t i = 0; i < a.size() && i < b.size(); i++) {
             res += a[i] * b[i];
         }
-    
         return res;
     }
 
-    static double getCoverage(std::vector<double> diagonals) {
-        double res = 0;
+    static float getCoverage(std::vector<float> diagonals) {
+        float res = 0;
         for (size_t index = 0; index < diagonals.size(); index++) {
             res += std::log(diagonals[index]);
         }
@@ -54,7 +53,7 @@ class LazyKernelMatrix : public KernelMatrix {
     private:
     const BaseData &data;
 
-    std::vector<std::unordered_map<size_t, double>> kernelMatrix;
+    std::vector<std::unordered_map<size_t, float>> kernelMatrix;
     std::unique_ptr<RelevanceCalculator> calc;
 
     // Disable pass by value. This object is too large for pass by value to make sense implicitly.
@@ -68,7 +67,7 @@ class LazyKernelMatrix : public KernelMatrix {
 
     LazyKernelMatrix(const BaseData &data, std::unique_ptr<RelevanceCalculator> calc) 
     : 
-        kernelMatrix(data.totalRows(), std::unordered_map<size_t, double>()),
+        kernelMatrix(data.totalRows(), std::unordered_map<size_t, float>()),
         data(data),
         calc(move(calc))
     {}
@@ -77,11 +76,11 @@ class LazyKernelMatrix : public KernelMatrix {
         return this->data.totalRows();
     }
 
-    double get(size_t j, size_t i) {
+    float get(size_t j, size_t i) {
         const size_t forward_key = std::max(j, i);
         const size_t back_key = std::min(j, i);
         if (this->kernelMatrix[forward_key].find(back_key) == this->kernelMatrix[forward_key].end()) {
-            double score = calc->get(forward_key, back_key);
+            float score = calc->get(forward_key, back_key);
             this->kernelMatrix[forward_key].insert({back_key, score});
         }
 
@@ -91,7 +90,7 @@ class LazyKernelMatrix : public KernelMatrix {
 
 class NaiveKernelMatrix : public KernelMatrix {
     private:
-    std::vector<std::vector<double>> kernelMatrix;
+    std::vector<std::vector<float>> kernelMatrix;
 
     // Disable pass by value. This object is too large for pass by value to make sense implicitly.
     //  Use an explicit constructor to pass by value.
@@ -105,16 +104,16 @@ class NaiveKernelMatrix : public KernelMatrix {
     static std::unique_ptr<NaiveKernelMatrix> from(
         const BaseData &data, 
         std::unique_ptr<RelevanceCalculator> calc) {
-        std::vector<std::vector<double>> result(
+        std::vector<std::vector<float>> result(
             data.totalRows(), 
-            std::vector<double>(data.totalRows(), 0)
+            std::vector<float>(data.totalRows(), 0)
         );
 
         // TODO: Verify that this is parallel safe, from a glance this looks super dangerous
         #pragma omp parallel for
         for (size_t i = 0; i < data.totalRows(); i++) {
             for (size_t j = i; j < data.totalRows(); j++) {
-                double score = calc->get(i, j);
+                float score = calc->get(i, j);
                 result[j][i] = score;
                 result[i][j] = score;
             }
@@ -123,13 +122,13 @@ class NaiveKernelMatrix : public KernelMatrix {
         return std::unique_ptr<NaiveKernelMatrix>(new NaiveKernelMatrix(move(result)));
     }
 
-    NaiveKernelMatrix(std::vector<std::vector<double>> data) : kernelMatrix(move(data)) {}
+    NaiveKernelMatrix(std::vector<std::vector<float>> data) : kernelMatrix(move(data)) {}
 
     size_t size() {
         return this->kernelMatrix.size();
     }
 
-    double get(size_t j, size_t i) {
+    float get(size_t j, size_t i) {
         return this->kernelMatrix[j][i];
     }
 };
